Added painting 'Under the Great Wave off Kanagawa'
This is one of the most iconic images of modern art and the most famous print of Katsushika Hokusai. The ingenuity of the composition, depicting a huge wave about to crash down  upwon frail voats beneath while the cresting foam on top of the wave wraps around Mount Fuji has been very influential on generations of artists both in the East and the West. The energy it conveys echoes similar approaches to render the 'subime' of nature in 19th centery European artistic movements. A smaller wave surging in the forefront mirrors the shame of Mount Fujo, a recurrent feature in the series. This is a fine and early impression of this great design, with blue contours, the outline of the pink clouds still showing in the sky when these have often disappeared from most impressions of the print.

Source: Sotheby's <a href="https://www.sothebys.com/en/slideshows/20-stunning-views-of-japan?slide=katsushika-hokusai-under-the-great-wave-off-kanagawa-thirty-six-views-of-mount-fuji-estimate-hk-400-000-600-000">20 Stunning Views of Japan</a>. Licensing and Copyright unknown.